requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
build==1.2.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
deepdiff==8.0.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
effdet==0.4.1
emoji==2.12.1
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.112.2
fastjsonschema==2.20.0
filelock==3.15.4
filetype==1.2.0
flatbuffers==20181003210633
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
greenlet==3.0.3
groq==0.11.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface_hub==0.24.6
humanfriendly==10.0
idna==3.8
ImageHash==4.3.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
iopath==0.1.10
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.1.1
json5==0.9.25
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kubernetes==30.1.0
langchain==0.2.16
langchain-chroma==0.1.3
langchain-cli==0.0.30
langchain-community==0.2.16
langchain-core==0.2.38
langchain-groq==0.1.9
langchain-nomic==0.1.2
langchain-ollama==0.1.3
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langchain-unstructured==0.1.2
langchainhub==0.1.21
langdetect==1.0.9
langgraph==0.2.16
langgraph-checkpoint==1.0.8
langserve==0.2.3
langsmith==0.1.112
layoutparser==0.3.4
libcst==1.4.0
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multimethod==1.12
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.3
nltk==3.9.1
nomic==3.1.2
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.3.2
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.2
openai==1.43.0
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orderly-set==5.2.2
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandas-profiling==3.2.0
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
phik==0.12.4
pi_heif==0.18.0
pikepdf==9.2.1
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.1
posthog==3.6.3
prometheus-client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==1.10.13
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pypandoc==1.13
pyparsing==3.1.4
pypdf==4.3.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject-toml==0.0.10
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.1
PyWavelets==1.7.0
PyYAML==6.0.2
pyzmq==26.2.0
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.38.4
sympy==1.13.2
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.9
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.5
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typer==0.9.4
types-python-dateutil==2.9.0.20240821
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.9
unstructured-client==0.24.1
unstructured.pytesseract==0.3.13
unstructured_inference==0.7.36
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.23.2
uvloop==0.20.0
visions==0.7.4
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
widgetsnbextension==4.0.13
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.9
zipp==3.20.1
